Timer support and opcode fixes

+++ keypad.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include "SDL3/SDL.h"
 
-//returns the scancode of the key being pressed
-int keyPressed();
+//returns the hex value of the key being pressed
+uint8_t keyPressed();
 
 #endif
